% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_haz.R
\name{bm_out}
\alias{bm_out}
\title{Background mortality table}
\usage{
bm_out(table1, table2)
}
\arguments{
\item{table1}{US 2015 Life Table for female}

\item{table2}{US 2015 Life Table for male}
}
\value{
None
}
\description{
Background mortality table
}
