% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_s.R
\name{bm_table}
\alias{bm_table}
\title{Extract Survival Probabilities from Life Table}
\usage{
bm_table(years, table)
}
\arguments{
\item{years}{age}

\item{table}{life table}
}
\value{
None
}
\description{
Extract Survival Probabilities from Life Table
}
