% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmcure.R
\name{bmcure}
\alias{bmcure}
\title{Main function to estimate parameters and print out}
\usage{
bmcure(formula, cureform, link, data, table1, table2,
  na.action = na.omit, method = "parametric", dis = "weibull")
}
\arguments{
\item{formula}{A survival formula based on the Surv() function, including survival time, right censoring indicator and covariates.}

\item{cureform}{The covariates included in the cure link function.}

\item{link}{The type of cure rate link function, including "logit","probit", and "cloglog".}

\item{data}{The data set needed.}

\item{table1}{US 2015 Life Table for female.}

\item{table2}{US 2015 Life Table for male.}

\item{na.action}{The action to deal with NA/missing values.}

\item{method}{The method to estimate parameters, including "semiparametric","flexible" and "parametric".}

\item{dis}{If "parametric" is specified for method, then dis must be provided, such as "exp","weibull","llogis","gamma","gompertz","lnorm" and "gengamma".}
}
\value{
None
}
\description{
Main function to estimate parameters and print out
}
