% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breastla.R
\docType{data}
\name{breastla}
\alias{breastla}
\title{Breast Cancer Dataset for Louisiana}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\usage{
data(breastla)
}
\description{
Pseudo data from SEER 2000-2012 Louisiana female breast cancer,
for regional stage, with 6200 observations, mean age of 65.7.
}
\keyword{datasets}
