% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com_like.R
\name{com_like}
\alias{com_like}
\title{Complete Likelihood Function in Parametric Method}
\usage{
com_like(par, Time, Status, Z, X, dis, w, link, mort.h, mort.s, ii)
}
\arguments{
\item{par}{all the parameters in the model.}

\item{Time}{Survival time from data set.}

\item{Status}{Right censoring status: 1 is death, 0 is censored.}

\item{Z}{The covariates included in the cure linke function.}

\item{X}{The covaraites for latency survival function.}

\item{dis}{If "parametric" is specified for method, then dis must be provided, such as "exp","weibull","llogis","gamma","gompertz","lnorm" and "gengamma".}

\item{w}{unobserved cure status indicator.}

\item{link}{The type of cure rate link function, including "logit","probit", and "cloglog".}

\item{mort.h}{background hazard rate.}

\item{mort.s}{background survival rate.}

\item{ii}{the number of observation.}
}
\value{
None
}
\description{
Complete Likelihood Function in Parametric Method
}
