% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_em.R
\name{emp}
\alias{emp}
\title{Parametric EM Algorithm.}
\usage{
emp(Time, Status, age, Z, X, mort.s, mort.h, link = "logit",
  emmax = 100, eps = 1e-07, dis = dis)
}
\arguments{
\item{Time}{Survival time from data set.}

\item{Status}{Right censoring status: 1 is death, 0 is censored.}

\item{age}{The varaible of age.}

\item{Z}{The covariates for cure link function.}

\item{X}{The covaraites for latency survival function.}

\item{mort.s}{background survival rate from life table.}

\item{mort.h}{background hazard rate from life table.}

\item{link}{The type of cure rate link function, including "logit","probit", and "cloglog".}

\item{emmax}{The maximum iteration is 100.}

\item{eps}{The stopping criteria for convergence is 1e-7.}

\item{dis}{The parametric distribution such as "exp","weibull","llogis","gamma","gompertz","lnorm" and "gengamma".}
}
\value{
None
}
\description{
Parametric EM Algorithm.
}
