% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish1.R
\name{fish_func1}
\alias{fish_func1}
\title{Likelihood function for cure link function in variance estimation}
\usage{
fish_func1(b, Z, w, link, ii)
}
\arguments{
\item{b}{parameters in cure link function.}

\item{Z}{The covariates included in the cure link function.}

\item{w}{unobserved cure status indicator.}

\item{link}{The type of cure rate link function, including "logit","probit", and "cloglog".}

\item{ii}{the number of observation.}
}
\value{
None
}
\description{
Likelihood function for cure link function in variance estimation
}
