% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish2.R
\name{fish_func2}
\alias{fish_func2}
\title{Likelihood function for latency survival function in variance estimation}
\usage{
fish_func2(beta, s, th, Status, X, w, mort.h, mort.s, ii)
}
\arguments{
\item{beta}{The estimated parameters in latency survival function.}

\item{s}{The estimated baseline survival rate.}

\item{th}{The estiamted total baseline hazard rate.}

\item{Status}{Right censoring status: 1 is death, 0 is censored.}

\item{X}{The covaraites for latency survival function.}

\item{w}{unobserved cure status indicator.}

\item{mort.h}{background hazard rate from life table.}

\item{mort.s}{background survival rate from life table.}

\item{ii}{the number of observation.}
}
\value{
None
}
\description{
Likelihood function for latency survival function in variance estimation
}
