% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex.R
\name{flex.spline}
\alias{flex.spline}
\title{Flexible Parametric EM Algorithm Main Function.}
\usage{
flex.spline(Time, Status, Z, X, Mbigs, Ibigs, r, b, beta, w, mort.s,
  mort.h, link = "logit", emmax = 100, eps = 1e-10)
}
\arguments{
\item{Time}{Survival time from data set.}

\item{Status}{Right censoring status: 1 is death, 0 is censored.}

\item{Z}{The covariates for cure link function.}

\item{X}{The covaraites for latency survival function.}

\item{Mbigs}{M spline basis.}

\item{Ibigs}{I spline basis.}

\item{r}{Initial value of r.}

\item{b}{Initial value of b.}

\item{beta}{Initial value of beta.}

\item{w}{Initial value of w.}

\item{mort.s}{background survival rate from life table.}

\item{mort.h}{background hazard rate from life table.}

\item{link}{The type of cure rate link function, including "logit","probit", and "cloglog".}

\item{emmax}{The maximum iteration is 100.}

\item{eps}{The stopping criteria for convergence is 1e-7.}
}
\value{
None
}
\description{
Flexible Parametric EM Algorithm Main Function.
}
