% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knots_sel.R
\name{knot.fun}
\alias{knot.fun}
\title{Knots Selection in Flexible Parametric EM Algorithm.}
\usage{
knot.fun(Time, Status, Z, X, b, beta, w, mort.s, mort.h, link = "logit",
  emmax, eps, maxk)
}
\arguments{
\item{Time}{Survival time from data set.}

\item{Status}{Right censoring status: 1 is death, 0 is censored.}

\item{Z}{The covariates for cure link function.}

\item{X}{The covaraites for latency survival function.}

\item{b}{Initial value for b}

\item{beta}{Initial value for beta}

\item{w}{Initial value for w}

\item{mort.s}{background survival rate from life table.}

\item{mort.h}{background hazard rate from life table.}

\item{link}{The type of cure rate link function, including "logit","probit", and "cloglog".}

\item{emmax}{The maximum iteration is 100.}

\item{eps}{The stopping criteria for convergence is 1e-7.}

\item{maxk}{The maximum number of knots to search}
}
\value{
None
}
\description{
Knots Selection in Flexible Parametric EM Algorithm.
}
