% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_tt.R
\name{lifett}
\alias{lifett}
\title{Generate Weibull Parameters from Life Table}
\usage{
lifett(weipar, try)
}
\arguments{
\item{weipar}{weibull parameters}

\item{try}{the data set of survival time and survival probabilities}
}
\value{
None
}
\description{
Generate Weibull Parameters from Life Table
}
