% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_like.R
\name{like.com}
\alias{like.com}
\title{Complete Likelihood Function in Flexible Parametric.}
\usage{
like.com(est0r, link, X, Z, Mbigs, Ibigs, Status, w, mort.s, mort.h)
}
\arguments{
\item{est0r}{The parameters to be updated.}

\item{link}{Survival time from data set.}

\item{X}{The covariates for latency survival function.}

\item{Z}{The covariates fir cure function.}

\item{Mbigs}{M spline basis.}

\item{Ibigs}{I spline basis.}

\item{Status}{Right censoring status: 1 is death, 0 is censored.}

\item{w}{unobserved cure status indicator.}

\item{mort.s}{background survival rate from life table.}

\item{mort.h}{background hazard rate from life table.}
}
\value{
None
}
\description{
Complete Likelihood Function in Flexible Parametric.
}
