% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll1.R
\name{ll1}
\alias{ll1}
\title{Likelihood function for cure link function in EM algorithm}
\usage{
ll1(tryb, Z, w, link)
}
\arguments{
\item{tryb}{parameters in cure link function.}

\item{Z}{The covariates included in the cure link function.}

\item{w}{unobserved cure status indicator.}

\item{link}{The type of cure rate link function, including "logit","probit", and "cloglog".}
}
\value{
None
}
\description{
Likelihood function for cure link function in EM algorithm
}
