% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll1r.R
\name{ll1r}
\alias{ll1r}
\title{First Likelihood Function to Update b in Flexible Parametric.}
\usage{
ll1r(tryb, Z, w, link)
}
\arguments{
\item{tryb}{The value of b to be updated.}

\item{Z}{The covariates for cure link function.}

\item{w}{Unobserved cure status indicator.}

\item{link}{The type of cure rate link function, including "logit","probit", and "cloglog".}
}
\value{
None
}
\description{
First Likelihood Function to Update b in Flexible Parametric.
}
