% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll2r.R
\name{ll2r}
\alias{ll2r}
\title{Second Likelihood function for latency survival function in Flexible EM algorithm.}
\usage{
ll2r(x, Time, Status, X, Mbigs, Ibigs, w, mort.s, mort.h)
}
\arguments{
\item{x}{The parameters to be updated.}

\item{Time}{Survival time from data set.}

\item{Status}{Right censoring status: 1 is death, 0 is censored.}

\item{X}{The covaraites for latency survival function.}

\item{Mbigs}{M spline basis.}

\item{Ibigs}{I spline basis.}

\item{w}{unobserved cure status indicator.}

\item{mort.s}{background survival rate from life table.}

\item{mort.h}{background hazard rate from life table.}
}
\value{
None
}
\description{
Second Likelihood function for latency survival function in Flexible EM algorithm.
}
