% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lungla.R
\docType{data}
\name{lungla}
\alias{lungla}
\title{Lung Dataset for Louisiana}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\usage{
data(lungla)
}
\description{
Data from SEER 2019 experiment on gravitropism in
Arabidopsis, with data on 162 recombinant inbred lines (Ler x
Cvi). The outcome is the root tip angle (in degrees) at two-minute
increments over eight hours.
}
\keyword{datasets}
