% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIspline.R
\name{MIspline}
\alias{MIspline}
\title{Basis Splines in Flexible Parametric Method.}
\usage{
MIspline(x, order, knots)
}
\arguments{
\item{x}{a row vector (grid of survival time)}

\item{order}{degree of splines}

\item{knots}{a sequence of increasing points}
}
\value{
None
}
\description{
Basis Splines in Flexible Parametric Method.
}
