% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_var.R
\name{per_fun}
\alias{per_fun}
\title{Perturbation variance estimation}
\usage{
per_fun(b = b, Z = Z, link = link, Status = Status, w = w,
  beta = beta, s = s, th = th, X = X, mort.h = mort.h,
  mort.s = mort.s)
}
\arguments{
\item{b}{parameters in cure link function.}

\item{Z}{The covariates included in the cure link function.}

\item{link}{The type of cure rate link function, including "logit","probit", and "cloglog".}

\item{Status}{Right censoring status: 1 is death, 0 is censored.}

\item{w}{unobserved cure status indicator.}

\item{beta}{The estimated parameters in latency survival function.}

\item{s}{The estimated baseline survival rate.}

\item{th}{The estiamted total baseline hazard rate.}

\item{X}{The covaraites for latency survival function.}

\item{mort.h}{background hazard rate from life table.}

\item{mort.s}{background survival rate from life table.}
}
\value{
None
}
\description{
Perturbation variance estimation
}
