% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotbmcure.R
\name{plotbmcure}
\alias{plotbmcure}
\title{Plot the predicted survival function.}
\usage{
plotbmcure(object, type = "S", xlab = "Time",
  ylab = "Predicted Survival Probability", ...)
}
\arguments{
\item{object}{An object from the function predictbmcure.}

\item{type}{The type of plot, such as "S" or "l".}

\item{xlab}{The x-axis label.}

\item{ylab}{The y-axis label.}

\item{\dots}{Other plot arguments.}
}
\value{
None
}
\description{
Plot the predicted survival function.
}
