% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictbmcure.R
\name{predictbmcure}
\alias{predictbmcure}
\title{Predict survival probabilities.}
\usage{
predictbmcure(object, newX, newZ, type, ...)
}
\arguments{
\item{object}{An object from the function predictbmcure.}

\item{newX}{The covariates in cure link function.}

\item{newZ}{The covaraites in latency survival function.}

\item{type}{The data type can be "whole" for whole data set, "female" for subgroup of female or "male" for subgroup of male.}

\item{\dots}{Other arguments.}
}
\value{
None
}
\description{
Predict survival probabilities.
}
