% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printbmcure.R
\name{printbmcure}
\alias{printbmcure}
\title{Print out estimation.}
\usage{
printbmcure(x, ...)
}
\arguments{
\item{x}{An object from estimation in the function bmcure.}

\item{\dots}{Other plot arguments.}
}
\value{
None
}
\description{
Print out estimation.
}
