% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_base.R
\name{psurv}
\alias{psurv}
\title{Parametric baseline estimation.}
\usage{
psurv(Time, dis, dpar)
}
\arguments{
\item{Time}{Survival time from data set.}

\item{dis}{The parametric distribution such as "exp","weibull","llogis","gamma","gompertz","lnorm" and "gengamma".}

\item{dpar}{the parameters for different distributions.}
}
\value{
None
}
\description{
Parametric baseline estimation.
}
