% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_time.R
\name{residual_time}
\alias{residual_time}
\title{Summary of Survival Probabilities from Life Table}
\usage{
residual_time(table1, table2, sex, age, year, country)
}
\arguments{
\item{table1}{US 2015 Life Table for female}

\item{table2}{US 2015 Life Table for male}

\item{sex}{sex}

\item{age}{age}

\item{year}{the year of life table}

\item{country}{the countro of life table}
}
\value{
None
}
\description{
Summary of Survival Probabilities from Life Table
}
