% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{surv.func}
\alias{surv.func}
\title{Semiparametric baseline estimation}
\usage{
surv.func(Time, Status, age, X, beta, w, mort.s, mort.h)
}
\arguments{
\item{Time}{observed survival time}

\item{Status}{right censored status (1=death, 0=censored)}

\item{age}{the variable of age}

\item{X}{covariates in latency survival function}

\item{beta}{parameters for X}

\item{w}{unobserved cure indicator}

\item{mort.s}{background survival rate}

\item{mort.h}{background hazard rate}
}
\value{
None
}
\description{
Semiparametric baseline estimation
}
