% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USAfemale2015.R
\docType{data}
\name{USAmale2015}
\alias{USAmale2015}
\title{USA 2015 Male Life Table}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\usage{
data(USAmale2015)
}
\description{
Data from SEER 2019 experiment on gravitropism in
Arabidopsis, with data on 162 recombinant inbred lines (Ler x
Cvi). The outcome is the root tip angle (in degrees) at two-minute
increments over eight hours.
}
\keyword{datasets}
